+++ app/containers/SearchMap.js
@@ -1,17 +1,18 @@
 /*
  * @flow
  */
-import { connect } from 'react-redux'
-import SearchMapComponent from '../components/SearchMap'
-import { addCustomLocationFilter, addCustomTrendLocationFilter } from '../actions'
-import type { Dispatch } from '../utils/flowtype'
+import { connect } from 'react-redux';
+import SearchMapComponent from '../components/SearchMap';
+import { addCustomLocationFilter } from '../actions';
+import type { Dispatch } from '../utils/flowtype';
 
 const mapStateToProps = (state) => {
     return {
         sensors: state.sensors.data,
         updateSensors: state.sensors.available_sensors,
         drawn_sensors: state.sensors.draw_available_sensors,
-        selectedLocation: state.selectedSearch.locations.selected
+        selectedLocation: state.selectedSearch.locations.selected,
+        shapeCoordinates: state.sensors.shape_coordinates,
     }
 };